+++ projects/plugins/boost/app/modules/critical-css/class-critical-css.php
@@ -65,13 +65,6 @@ class Critical_CSS extends Module {
 	);
 
 	/**
-	 * True if this pageload is generating Critical CSS.
-	 *
-	 * @var bool
-	 */
-	public $generating_critical_css;
-
-	/**
 	 * List of all the Critical CSS Types.
 	 *
 	 * The order is important because searching for critical CSS will stop as soon as a value is found.
@@ -152,10 +145,7 @@ class Critical_CSS extends Module {
 			add_action( 'wp_footer', array( $this, 'onload_flip_stylesheets' ) );
 		}
 
-		// Check for the GET parameter indicating this is rendering for CSS generation.
-		$this->generating_critical_css = $this->check_generate_query();
-
-		if ( $this->generating_critical_css ) {
+		if ( $this->is_generating_critical_css() ) {
 			add_action( 'wp_head', array( $this, 'display_generate_meta' ), 0 );
 			$this->force_logged_out_render();
 		}
@@ -450,11 +440,18 @@ class Critical_CSS extends Module {
 	}
 
 	/**
-	 * Check for GET parameters or Headers indicating the current request is
-	 * generating Critical CSS.
+	 * Returns true if this pageload is generating Critical CSS, based on GET
+	 * parameters and headers.
+	 *
 	 * phpcs:disable WordPress.Security.NonceVerification.Recommended
 	 */
-	public function check_generate_query() {
+	public function is_generating_critical_css() {
+		static $is_generating = null;
+		if ( null !== $is_generating ) {
+			return $is_generating;
+		}
+
+		// Accept nonce via HTTP headers or GET parameters.
 		$generate_nonce = null;
 		if ( ! empty( $_GET[ self::GENERATE_QUERY_ACTION ] ) ) {
 			$generate_nonce = sanitize_key(
@@ -466,15 +463,15 @@ class Critical_CSS extends Module {
 			);
 		}
 
-		if ( empty( $generate_nonce ) ) {
-			return false;
-		}
+		// If GET parameter or header set, we are trying to generate.
+		$is_generating = ! empty( $generate_nonce );
 
-		if ( ! Nonce::verify( $generate_nonce, self::GENERATE_QUERY_ACTION ) ) {
+		// Die if the nonce is invalid.
+		if ( $is_generating && ! Nonce::verify( $generate_nonce, self::GENERATE_QUERY_ACTION ) ) {
 			die();
 		}
 
-		return true;
+		return $is_generating;
 	}
 	// phpcs:enable WordPress.Security.NonceVerification.Recommended
 
@@ -710,7 +707,7 @@ class Critical_CSS extends Module {
 	 */
 	public function should_display_critical_css() {
 		// Don't display Critical CSS when generating Critical CSS.
-		if ( $this->generating_critical_css ) {
+		if ( $this->is_generating_critical_css() ) {
 			return false;
 		}