+++ projects/packages/connection/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.32.0] - 2021-11-30
+### Added
+- Added a way to set the has_seen_wc_connection_modal option from the API
+- Provides an Initial State that can be used by JS packages
+
+### Changed
+- Updated package dependencies.
+
 ## [1.31.0] - 2021-11-22
 ### Added
 - Added plugin_slug parameter to the v4/register endpoint
@@ -470,6 +478,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Separate the connection library into its own package.
 
+[1.32.0]: https://github.com/Automattic/jetpack-connection/compare/v1.31.0...v1.32.0
 [1.31.0]: https://github.com/Automattic/jetpack-connection/compare/v1.30.13...v1.31.0
 [1.30.13]: https://github.com/Automattic/jetpack-connection/compare/v1.30.12...v1.30.13
 [1.30.12]: https://github.com/Automattic/jetpack-connection/compare/v1.30.11...v1.30.12